innovation-lab-examples/llama-index/requirements.txt
llama-index-core==0.14.15
llama-index-llms-openai==0.6.18
llama-index-embeddings-openai==0.5.1
llama-index-vector-stores-qdrant==0.9.1
llama-index-readers-file==0.5.6
qdrant-client==1.17.0
uagents==0.23.6
uagents-core==0.4.2
python-dotenv==1.2.1
stripe>=7.0.0
requests==2.32.5
openai==2.21.0
